feat: Restrict loading of script attributes based on nonces and origins
Only scripts from a known origin and having a nonce can now be loaded.
This achieved by setting 2 CSP in the header, one is responsible to
restrict the origins the other one to ensure the presence of a nonce.
Browsers apply restrictions in a cumulative way.

The policy is still quite weak since we still allow unsafe script
attributes.

Additionally, the variables used to set the CSP header in pre.php do not
pollute the global scope anymore.

Part of request #17967: Deploy a useful content security policy

Change-Id: I08db1e195ef707e367df61afdfcd308bcc89f52f