feat: SVG favicon should be last
part of story #45211: Tuleap 17.0 new theme

How to test:
- If you load any Tuleap page with Firefox, with dev tools open and
without cache, you should see it download the SVG favicon
- Same for Chromium

Why?
The order of the declaration matters, the last one "wins". Browsers that
support SVG icons (all our target browsers) will use the SVG one if it
is declared last. Browsers that don't support them will fall back to
PNG, and if not that to ICO.

I also tried to use full paths in the web manifest files. I'm not
convinced those are actually useful, but
https://realfavicongenerator.net/favicon-checker reported that those
icons were not found.

Change-Id: I80d25b705fbdbc7fe9b387d2483c1d89105d3ade