feat: SVG favicon is square, grey favicon adapts to dark theme
part of story #45211: Tuleap 17.0 new theme

How to test:
- Force reload your dev page (clear cache, for example with dev tools)
- Enable the feature flag:
tuleap config-set feature_flag_allow_favicon_variant 1

- Check the favicon theme color checkbox in your user preferences.
- When your user has Grey color theme and you set your browser or OS to
Dark mode, the SVG favicon should have a lighter color so that it
remains readable on a browser tab with dark background.

Why?
Favicons should be square, it bothered me that our SVG favicon was the
only format that was not square (19x20 px). I used Inkscape to scale it
to 32x32 which is multiple of 8, and to center the path on the image.

I dropped the <style> tags for all colors except grey where it is
replaced to adapt to the dark theme. As far as I could tell, they did
nothing: filter:none is the default value, so both in Light and Dark
theme, it would reset the filter to the default.

I used https://colordesigner.io/color-palette-builder to select a Tint
variation for the grey theme color and landed on #aaaeb0. It should be
readable enough in Dark theme, while still being Grey, not white.

I used SVGO (https://svgomg.net/) with default options + "Style to
attributes" and Multipass enabled. The image size: 4 KiB -> 2.3 KiB.

Change-Id: Ic225e05263b31eb29f422ae9579dd562638bd444