refactor(cross-tracker-search): use data-model to build result table
part of story #44821 Include linked artifacts to XLSX export

This change introduces the use of the data-model into the
SelectableTable stack in order to base the building of the view off the
data-model instead of building the data-model alongside the view.

Do note that during the refactoring we took the liberty to remove two
things that were inducing more code for no real value. We can
re-introduce them afterwards if needed:

1. There is no more spinner in the "search" button in the Edit & Create
   views: when you click on the button it is already disabled and there
   is already a much bigger spinner on your screen in place of the result
   table.
2. There used to be a variant of the "No artifact found" screen if there
   was no query saved in the widget that was inviting you to create a
   query: problem is, the only way to see this screen is to precisely
   NOT have permissions to create a new query, because if you can the
   screen is directly replaced by the Query Creation interface, so this
   was at best misleading. Now you only see a "No artifact found"
   message but in any case as a simple user there is nothing you can do
   about it.

Co-authored-by: Marie-Ange Garnier <marie-ange.garnier@enalean.com>
Change-Id: I919b864db44e28ad3c2c17278d23663bf7259583