fix: Use primary button when no color for provider
When an OIDC provider does not have a color, then the button appears
broken on the login page (the button is half styled).

Now if no color, the button is primary (consistency with other primary
buttons on the page). This is usefull when the OIDC is the only way to
connect to the platform and the button is unique on the page.

The benefit is that we can provide Tuleap orange buttons on our default
platforms (tuleap.net, tuleap-devel, …) instead of a Clokcwork orange.

Note: Clockwork orange is not changed to Tuleap orange because users of
TLP Swatch Colors should not be impacted by Tuleap rebranding.
Furthermore it would imply to adjust other colors to bring harmony that
would be broken by changing existing orange (the new orange has a more
reddish hue).
Likewise, we cannot add Tuleap orange to the TLP Swatch Colors: it will
produce a even numbered palette, and some colors have 3 variations, why
orange would have 4 variations?

Part of story #45211: Tuleap 17.0 new theme

Change-Id: I48a41ab02367310be732d827871dfb68c67f03f6