feat: PNG default logo is ignored again
part of story #45211: Tuleap 17.0 new theme

How to test:
- run "make bash-web"
- backup the logo file at /var/lib/tuleap/images/organization_logo.png
- copy the logo file from
/usr/share/tuleap/src/www/themes/BurningParrot/images/organization_logo.png
to the /var/lib folder
- Run "tuleap -c"
- Go to a page that scrolls, for example a user profile, or a dashboard.
Before this patch, you should see the PNG logo (when it scrolls, it
gains the theme background).
- After this patch (clear the cache again), you should keep seeing the
SVG logo (the text changes color depending on the background).
- You can restore the backed-up logo.

Why?
During
git #tuleap/stable/eabd2109fe9b902e609c3678cbb2751e9cbaff38,
we introduced two new PNG default platform logos: one for common theme
and one for BurningParrot theme folders. The hash to detect it
(and ignore it in favor of the SVG default platform logo) was changed,
but we only added the one for the common theme logo. When installing a
new platform, the BurningParrot one takes precedence, and since its hash
is not in the list, Tuleap wrongly decides that there is a customized
PNG logo. Unit tests are changed to indicate to future maintainers that
the hash for the default logo should be added to the list if the image
changes.

The problem was hard to reproduce, because on older platforms the older
PNG image was deployed in the /var/lib/tuleap/images/ folder. When a
file is present, the "tuleap-cfg site-deploy:images" command will not
erase it (because it means it was customized). This is why the problem
was visible on Cypress (the platform is reinstalled every time) and on
some newly installed platforms.

Change-Id: I3e95ae15ed9a0738ea2d863fb1ba24a5c159b663