feat: Publish the favicon variant preference for MediaWiki
part of story #45211: Tuleap 17.0 new theme

How to test:
- Enable the feature flag:
tuleap config-set feature_flag_allow_favicon_variant 1
- Check the favicon theme color checkbox in your user preferences
- In the API explorer, the GET /projects/{id}/3rd_party_integration_data
REST route should return a new property: should_display_favicon_variant
It should be true with the previous steps. If you uncheck the
preference, or if you disable the feature flag, it should be false.

Why?
This is necessary to have a consistent favicon display on MediaWiki
Standalone service. When it is false, the favicon will be always orange,
otherwise it will adapt to the color theme.

Change-Id: I9a38e412c51e7558afa92a551bd4c2c9352b5767