chore: Convert FRS to BurningParrot
No functional changes… well… not too many.

TL;DR: The existing FRS layout is difficult to convert to TLP, so this
change will probably break some users' workflow.

(╯・ω・)╯︵ ┻━sɹɟ ʎɔɐƃǝl━┻

From both a UI and UX point of view, it was too complicated to convert
the existing FRS front page to TLP without changing the workflow.
Nowhere else in Tuleap do we have such nested table containers that can
expand inside each others on the same page: packages » releases » files.

Personal opinion: ideally we should probably ditch the FRS entirely and
let the users upload their files in Document service. That way they
would have full control over the hierarchy.  But that is a different
direction and would have more business consequences than "just" a
BurningParrot conversion.

So instead of trying to having a messy mix of tlp-pane and tlp-table on
the same page, the current proposal is to do it the Tuleap way:
* A page that list the packages (where you can create/edit/delete them)
* Which leads to a page that lists the releases of a package (where you
  can create/edit/delete them)
* Which leads to a page that lists the files of a release

This is simpler. The eventual regression is that you can no longer
easily compare the contents of two releases: you will have to open them
in two separate browser tabs instead of expanding them on the same page.
I'm not sure if this behavior is really desired; if not, this will at
least allow us to raise a proper business case instead of making
assumptions.

Please note that there are two different pages listing the files of a
release: the default one, and the one provided by the `frs` plugin.
Maybe we should adapt the plugin layout, but I am not sure that is
really in the scope of the current request.

This changes allows us to remove all inline JavaScript that was used to
confirm deletion (now replaced by modals), to select users (replaced by
lazybox), or to manage release creation/update (date pickers, adding
files, etc.). I also took the opportunity to simplify some dynamic
behavior (e.g.  adding a file to a release) or to fix some bugs (e.g.
adding comments to files picked from the staging area).

Note: to test the staging area while adding file to a release, you can
create `ftpadmin:ftpadmin` files in `/var/lib/tulaep/ftp/incoming` (This
folder contains files uploaded by ftp and/or files uploaded via web UI
that failed to be moved to the final directory for any reason.)

One drawback of having a nice UI framework like TLP is that when the
interface starts going crazy, layout issues become very visible. For
example, the files table in release edition mode contains many column,
most of which have a form element. This forces the table to have a very
large width, so the viewport has to scroll. It was fine in the legacy
interface because there was no visible structure, but now the large
table overflows its parent tlp-pane. To fix this, the same behavior as
in the document search is reused: the first cell is sticky, and the
table scrolls inside the tlp-pane. Yay consistency! \o/

I think that adding a breadcrumb would make the navigation between
packages and releases easier. If we think that this is needed, we should
probably do it later since this contribution is already quite huge. The
entanglement of various FRS legacy quirks and the need to rework the UX
make it difficult to split this into smaller batches. Fellow reviewer,
please accept my sincere apologies for the burden I'm putting on you.

Closes request #39758: Get Rid of Flaming Parrot in Files

Change-Id: Iaf7bf754815ecddf4523105467cf1a7c05d86da0