chore: Move dev/build env to Node 24
No functional changes. Everything still builds.

Needed because we need to bump the underlying npm
to a more recent versions to support publishing using the new "Trusted
Publisher" feature of npmjs.com.

Part of request #45235 Rework the publication of our npm packages to deal with the new npmjs.com rules

Change-Id: I704f39ddaabc9f692dc586be8fc8d152330d336b