chore: convert "forkRepositories" view to BP
part of request #39761 Get rid of Flaming Parrot in Git service

In this contribution, only the first step of the repository fork UI is
migrated to BurningParrot. The second and third steps (Permissions
selection and fork actions) will be migrated in the next contributions.

Why is this patch so big?

In order to get out of GitViews, I had to define a FrontRouter route
/projects/<project-unix-name>/fork-repositories. It was necessary
because we could not just make GitPluginDefaultController implement
DispatchableWithBurningParrot, otherwise all the other administration
views would have been broken. Having it managed into a dedicated
controller allowed me to migrate the view in a cleaner way, and with a
more convenient assets management.

The view was originally generated by echoing html tags. It is now
generated using dedicated presenters and mustache templates. The
presenters being built by a presenter builder, it had to be unit tested.

The frontend part of the UI was written in JQuery and was using codendi
globals from the browser (see removed code in git.js). It was not
working in a BurningParrot environment, and had to be rewritten
entirely in TypeScript with its own Vite config. I've also written unit
tests for it, which did not exist before.

How to test:

1. UX
- Go to the Git service of a project where there is no repository
- In the breadcrumb, click [Fork repositories]
--> A tlp-alert says "There is no forkable repositories in this project"

- In a project having git repositories, click [Fork repositories]
--> The list of forkable repositories is displayed as a multiple select
--> The "personal fork" option is selected by default
--> The project select box is disabled
--> The submit button is disabled
--> You can define a custom fork path

- Select repositories, define a custom path
--> The submit button is not disabled anymore
--> The preview updates

- Click the "project fork" option
--> The project selectbox is now enabled
--> The "custom fork path" inpout is now disabled
--> The preview has updated

2. Personal fork
- Select repositories
- click the "personal fork" option
- [optional] Set a custom fork path
- Submit

--> You land on the "Set fork permissions" page (still in FlamingParrot)
--> It sums up which repositories will be created and where

- Set the permissions and submit

--> You land on the project's repository list page
--> A feedback info says "Fork created successfully"

3. Cross project forks
- Select repositories
- click the "project fork" option
- Submit

--> You land on the "Set fork permissions" page (still in FlamingParrot)
--> It sums up which repositories will be created and where

- Set the permissions and submit

--> You land on the destination project's repository list page
--> A feedback info says "Fork created successfully"

4. Error management
Whenever an error occurres during the creation flow, you are supposed
 to be redirected to the fork repositories page AND have a feedback
 error message. The easiest way to test it is as follow:

- Select repositories
- click the "personal fork" option
- Set a custom path containing '..'
- Submit
- Submit the permissions
--> You get redirected to the fork repositories page
--> There is a feedback error

Change-Id: Ic7987e87daa990ec40a305c0b2413e4d5af7860d