refactor: Move BP global admin tabs to Git plugin
part of request #39761 Get rid of Flaming Parrot in Git service

How to test:
No functional change expected in Git > Administration > GitLab Group
Link tab:
- When there is no linked group, it renders the admin tabs as before and
shows a Vue wizard (see art #29251 for details).
- After linking a group, it shows the group settings. The admin tabs are
shown as before.
- It should work as before with or without project background image.

Why?
Before converting the other global admin tabs to BurningParrot, we need
to have a way to render the administration navigation tabs. To do so, we
move the mustache and presenter to the Git plugin. The CSS is also moved
to the Git plugin.

In the next steps, we will be able to convert the other tabs to
BurningParrot and reuse the code.

Change-Id: I8d0cfaa977068e5c56443e46dbef86f0e0dc3218