refactor: Convert TLP color picker to BurningParrot
Part of request #39759 Get Rid of Flaming Parrot in Trackers

The previous component @tuleap/plugin-tracker-color-picker was using
FlammingParrot styles. In this patch the component is re-written with
BurningParrot theme using a popover instead of a dropdown.

The component is available in the storybook

*No functional changes expected*

In tracker fields administration and planning adminstration, the
component should continue to work as before

Change-Id: I4a5ba35051ee7d03dd8f12906c46797ed4329720