chore: request #45254 Go: 1.25.1 -> 1.25.3
Our derivation to build and/or to get an dev env have been moved
to the go_latest package. The go_latest is not used within nixpkgs
making things easy if we want to overlay it without dealing with
rebuilding other tools of our stack. It is also more aligned with our
expectations for building our own package.

Go 1.25.3 is pulled from nixpkgs 25.05 branch, it has not yet reached
the "unstable"/rolling release branch since it is a mass rebuild there.

Change-Id: I397d3d64041d48d7f896ddc0a9b89ade6ea50c71