feat: Support inverted color for SVG logo
Tuleap supports [organization logo in SVG format][0]. Using SVG allows
to adapt the colors of the logo depending on the context, by using
dedicated CSS classnames:
* `logo-svg-fill-theme-color`: fill with the current theme variation
* `logo-svg-fill-default-text-color`: fill with the current text color

This contribution is a complement to `logo-svg-fill-default-text-color`.

When the logo is displayed on the homepage, user profile (after
scrolling), or user dashboard (after scrolling), then the background is
light (white in most of the cases).  When the logo is displayed on user
profile (without scrolling), user dashboard (without scrolling),
siteadmin, or in the context of a project, the background is dark
(background of the sidebar).

Using the aforementionned classname allows the SVG to adapt the color:
on light background `logo-svg-fill-default-text-color` will fill in dark
color, on dark background it will fill in white.

Everything works perfectly until you try to enclose your text in a dark
element in your logo. For example you might want that the text "enterprise"
is on a dark color badge.

+--------------------+
|                    | <----- SVG (transparent background)
|    ┏━━━━━━━━━━━━┓  |
| ☺  ┃ enterprise ┃<--------- Badge (white text on dark background)
|    ┗━━━━━━━━━━━━┛  |
|                    |
+--------------------+

Issue: when this logo is displayed on a dark background you wont see
anymore the badge (dark background on dark background). You need to
invert the colors to have a badge with a dark text on white background.

We can use the classname `logo-svg-fill-default-text-color` for that
purpose. The background of the badge will be white on dark background
and dark on white background. But what about the text itself? We need an
"invert" class in order to be able to fully invert the badge.

So starting from now, SVG will have both
`logo-svg-fill-default-text-color` and
`logo-svg-fill-default-text-color-invert` to manage this situation.

You can try with the following beautiful logo: save it as
`/var/lib/tuleap/images/organization_logo.svg` and clear the cache
`tuleap -c`.

```
<svg xmlns="http://www.w3.org/2000/svg" width="40" height="40" viewBox="0 0 10.583 10.583">
  <g transform="translate(-91.446 -96.123)">
    <circle cx="96.738" cy="101.414" r="5.292" fill="#1b1512" class="logo-svg-fill-default-text-color"/>
    <path d="m373.138 411.356-23.505-12.507-23.629 12.273 4.632-26.22-18.974-18.679 26.367-3.698 11.902-23.817 11.665 23.934 26.33 3.96-19.159 18.49z" fill="#fff" class="logo-svg-fill-default-text-color-invert" transform="matrix(.06348 0 0 .06348 74.532 77.606)"/>
  </g>
</svg>
```

You should see a star in a circle ✪ (maybe a dragon ball? 🐉) on every
pages (with and without scroll): homepage, user profile, user dashboard,
project pages, siteadmin.

Without this contribution, you cannot see the star on some pages.

Part of story #45211: Tuleap 17.0 new theme

[0]: https://docs.tuleap.org/administration-guide/system-administration/customisations.html

Change-Id: Ic689827c1b13214af7077bdffd232472d2593150