refactor: Extract document details tab in its own component
Part of request #39756 Get rid of Flaming Parrot in Document

With all the work on document, we add several tabs in the document
details. Each time we add one tab, we need to touch to every
Display*.vue to add the tab.

By extracting them in one component we'll have only this one to update.

*No functional changes expected*

Tabs should still be displayed as before and work

Change-Id: I419275cb125e2072d1d552f3f7d5c4bea8efcaba