feat: Improve XTS assistant pre-prompt
Try to better guide LLM into generation of a correct query.

This patch also fix a bug with permission verification that prevented
fields to be given as pre-prompt.

I also let more time for mistral to answer because when discussions
start to be long, 30s is too short.

Part of: story #44884 TQL assistant prototype

Change-Id: If61c369e0ded7467106519cd89190621ccd0466b