fix: Unnamed FRS package
For unknown reason, one can create a package without name. With the new
interface, the table list an empty entry and the user cannot access the release of this package.

Instead of silently allowing this, now the row is clearly identified as
problematic, and users cannot create anymore unnamed packages.

Closes request #45584: Unnamed FRS package

Change-Id: Ic2b73ad736394c7d6a557c3a120750e1148da76f