chore: convert "forkRepositoriesPermissions" view to BP
part of request #39761 Get rid of Flaming Parrot in Git service

Note 1:

Since the Access control form is shared between the fork repositories
permissions and the repositories permissions admin, I had to extract the
build of the access control presenter from GitForkPermissionsManager
into a dedicated presenter builder, then call it wherever needed. In
order not to "break" the UI in the repository permissions admin, I've
temporarily duplicated the mustache templates. They will be removed for
good when the permissions tab will be migrated to BurningParrot.

Note 2:

When forking repositories with defined fine grained permissions, all the
fields are readonly, meaning we don't need to migrate the frontend yet
(see `plugins/git/scripts/legacy/src/permissions.js`). The scripts will
be rewritten in TS in a dedicated commit.

How to test:
1. Personal fork, no fine-grained permissions
- Select one or more repositories to fork who do not have fine-grained
  permissions
- [optional] Set a custom path
- Submit
--> You land on the repositories permissions page.
--> It summarises which repositories will be forked and where.
--> You can set the read/write/rewind permissions

- Submit
--> The repositories have been successfully forked

2. Cross projects fork, no fine-grained permissions
- Select one or more repositories to fork who do not have fine-grained
  permissions
- Select the "project fork" option and a destination project
- Submit
--> You land on the repositories permissions page.
--> It summarises which repositories will be forked and where.
--> You can set the read/write/rewind permissions

- Submit
--> The repositories have been successfully forked

3. With fine grained permissions
- Select one or more repositories to fork who have fine-grained
  permissions
- Submit
--> All the permissions are displayed and are readonly

- Submit
--> Success

3. Regressions
- No regression expected in git repository admin > permissions and
  global admin > git templates

Change-Id: I36aaba357504000357e77810ce7bee2ae67ab680