feat: convert Git gerrit templates pane to BurningParrot
Part of request #39761 Get rid of Flaming Parrot in Git service

How to test:
- As Git administrator or Project administrator, go to the Git global
admin.
- You should still be able to do all the actions:
  - Create a new Gerrit template from scratch
  - Create a new template based on an existing template
- Create a new Gerrit template from a Gerrit server (you can fake it
by editing \Git_Driver_Gerrit_ProjectCreator::getGerritConfig(), you can
make it return a string)
  - Edit an existing template
  - Delete a template
- View as read-only a parent template (I edited the presenter to reuse
the project's templates)

The contents of the Gerrit templates should be fetched asynchronously.
Now there should be spinners indicating that something is loading. In
case of error, there should be a feedback either in the "Create" modal
or at the top of the table.

Why?
Previously, there was only one form with many parts shown/hidden
depending on the action. It was pretty hard to get a mental model of
what is shown for each action. I chose to split it into 5 different
modals. It does introduce repetition of the modal layout, but in my
opinion it's easier to understand now.

Change-Id: If0cec6244978730e238fae5328b6b75bed0cf5ab