feat: Build Sofware-Bill-of-Materials of PHP dependency trees
These SBOMs will later on be used to check the licenses used by
dependencies against an allow-list.

The SBOMs are generated in the CycloneDX format. It looks like
it will be easier for us to handle the CycloneDX than SPDX with
the current ecosystem used in Tuleap.

To test you can run `make generate-sbom`, you will SBOM alongside the
Composer files.

Part request #45595 Allowlist licenses acceptable for Tuleap dependencies

Change-Id: I96863c1a43450829f7200d57c94897d31a564414