chore: do_for_repositories in a PSR15Compatible controller
part of request #39761 Get rid of Flaming Parrot in Git service

This is a bonus patch. The goal is to completely move the code dealing
with forks outside of the MVC pattern carried out by Git/GitViews/GitActions.

Note:

I took the opportunity to add the personal fork path check to the step 2
in order not to wait until we reach the step 3 to get the "path cannot
countain double dots" error.

How to test:
- Fork a personal repository with no custom path
--> Success feedback, repository created under the u/<username> namespace,
redirected to repo list.

- Fork a personal repository with a custom path
--> Success feedback, repository created under the u/<username>/<path>
namespace.

- Fork a personal repository with a path containing double dots (e.g
  '../my-path')
--> Error feedback, repository not created, redirected to step 1

- Fork a personal repository already existing in the target
--> Warning feedback, repository not created, redirected to repo list

- Do a cross project fork with repositories not existing in the target
--> Success feedback, repository created, redirected to repo list

- Do a cross project fork with repositories alreay existing in the
  target project.
--> Warning feedbcak, repositories not created, redirected to
destination project repo list.

Change-Id: I9c24b86b832a242bffd4cd35295c1aa4d923a8e3