chore: Convert SVN notifications to BurningParrot
In order to simplify the conversion and the user interface, the
user-and-ugroup autocompleter has been removed in favor of three
different inputs: one for user groups, one for users, one for emails.

Furthermore the interface for managing notifications now looks like the
rest of Tuleap: a table listing notifications, table action to add a new
one, and edit/delete buttons for existing ones.

This was the last remaining FP page of the plugin.

Note: if you want to compare with the existing behavior, master is
broken due to 7a85b3c14a69daed40ec1aad1523d9f0e435af13, you need to
checkout Tuleap 17.0.99.34 or before.

Closes request #39757: Get Rid of Flaming Parrot in SVN Plugin

Change-Id: I12848c50e4bee3133f7626ae2030aac4fd58eff3