chore: Convert Tracker Global Rules to BurningParrot
The creation and deletion of rules is now consistent with the rest of
Tuleap: a table to list the existing rules, a table action to add a new
one (via a modal), and a button action to delete an existing one (with
confirmation modal).

In order to ease the conversion, the ability to update an existing rule
has been removed. Users will have to delete and add a new one. The
process to add a rule is simple and let us consider this regression good
enough.

Part of request #39759: Get Rid of Flaming Parrot in Trackers

Change-Id: I0ac73a5a006304254b369d9cef08dabf35b3d5d3