feat: request #45620 Upgrade from MW 1.39 to 1.43
Related changes in the extensions/skins:
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/TuleapWikiFarm/+/1205108
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/TuleapIntegration/+/1205114
https://gerrit.wikimedia.org/r/c/mediawiki/skins/TuleapSkin/+/1205149
https://gerrit.wikimedia.org/r/c/mediawiki/skins/TuleapSkin/+/1023451
https://gerrit.wikimedia.org/r/c/mediawiki/skins/TuleapSkin/+/1205155

Release announcements:
https://www.pro.wiki/articles/mediawiki-140-is-new
https://www.pro.wiki/articles/mediawiki-141-is-new
https://www.pro.wiki/news/whats-new-in-mediawiki-1-42-release
https://www.pro.wiki/news/what-new-features-in-mediawiki-1-43-release

To test rebuild the package, re-deploy it on your dev instance and do
not forget to run `tuleap-cfg site-deploy` to run the migration steps.

The tuleap-mathoid package has been removed, it is deprecated upstream
and the math extension can now leverage native browser features to do
the rendering.

Change-Id: I4c51e9653d5654946f65756fa2f3527cff5d653a