chore: remove legacy services inheritance
Neither 'svn' (SVN Core) nor 'tracker' (Tracker v3) are supposed to be
duplicated at project creation, whether it's XML or internal
duplication.

This commit removes all the legacy mechanisms for potential duplication
and ensures it's not possible to start a new project with one of the
legacy service.

Part of: request #16583 Tracker V3 Removal
Part of: request #29976 Remove SVN core

Change-Id: Ibff57feecde1d9351c5d2f1d380ced71e3582c2c