feat: Ensure JS manifest files are consistent with the built assets
With this change we should now be able to catch issues in the build
pipeline where top level are present in the manifest but are missing
in the built assets.

This should allow to catch issues such as 967942687c350fbe9abf14853b6d80ab804c2a5f
early.

Part of request #45595 Allowlist licenses acceptable for Tuleap dependencies

Change-Id: Ic0e9ad377d374f9a274a2f00a24be9f1ff4a66fe