chore: Convert Triggers to BurningParrot
No functional changes

My initial objective was to deprecate the existing implementation and
rebuild the user experience to align with the BurningParrot standards
(tlp-table with top-level action buttons, creation modal, etc.).
However, given the functional complexity of the feature, the end-to-end
effort – from design to implementation – would be too significant (it
would likely require its own dedicated Vue application). As a result,
the existing behavior is preserved and merely migrated to TypeScript,
removing dependencies on the tuleap.tracker global namespace as well as
on PrototypeJS/Scriptaculous.

Part of request #39759: Get Rid of Flaming Parrot in Trackers

Change-Id: I56e66312a12c2566ce975a5dea1d6d363dd422af