chore: Remove usage of the clean-webpack-plugin
This feature is now available natively in Webpack.
This is take 2 of 66fa0357895910803d2721c9b7754114d603cef7.
Cleaning of the @tuleap/core package is disabled to avoid issues,
the package is considered legacy anyway and has a manual cleanup
in the build script.

Part of request #45595 Allowlist licenses acceptable for Tuleap dependencies

Change-Id: I8642db4f270331df99139a76e2ec763d073cfe67