fix: load the hierarchy only on existing items
Part of request #39756 Get rid of Flaming Parrot in Document

Why?
The tabs in the new UI display an error when attempting to access logs,
referecnes or notifications for the root folder.
This is due to the load of the hierechy for the parent item. But as the
root folder has no parent, this causes an error.

How to test:
-- In the document service, click the ellipsis button
-- Click on 'Logs'
=> It should display the logs in the new UI

-- Return in the document service, click on the ellipsis button
-- Click on 'References'
=> It should display the references in the new UI

Change-Id: Ic4b31696e4c895b01522a0fc923829df6afbe1ce