fix: fine-grained-permissions template not loaded when forking
part of request #39761 Get rid of Flaming Parrot in Git service

I don't know why it happens, but I guess loading the same partial from
different locations did not help. To fix this issue, I've moved
everything concerning the access control form in an access-control/
folder and added it to the list of templates directories (otherwise it
won't load).

How to test:
- Fork a repository having fine grained permissions
--> The fine grained permissions are shown during step 2

- Define the permissions of git templates
--> The fine grained permissions form is shown

- Go to the permissions admin of a repository having fine grained
  permissions
--> The fine grained permissions are shown

Change-Id: I7ec497772fb497cfa24b3a465b9890d33dc136ff