chore: remove TV3/TV5 dual reference handling
No functional changes for tv5 cross references.

Functional changes for tv3: they are still extracted but the underlying
artifact is not reachable.
This closes request #45642: Reference on artifacts should not check artifact existence

On project administration » reference, the tv5 nature is renamed to
"Artifact".
Legacy natures are not anymore displayed in the table (artifact, news,
forum, cvs)

Part of request #16583 Tracker V3 Removal

Change-Id: I5ea3235e48aba41ab2af70e61f44968ccb6cab6c