feat: Add no approval table state
Part of request #39756 Get rid of Flaming Parrot in Document

When the document does not have an approval table, we tell it. If user
have write permission on the document there is also a button redirecting
to old ui for table creation

*Testing:*

With a document which can have an approval table but have none yet, go
to the approval table tab.

If you have write permission, create the table. After creation, in the
old ui you have an alert info with a link to the app.

You should now see the warning alert "Under development"

Change-Id: I4d81c1bb11559135f2e1704f1084b8a07e631160