feat: introduce error|empty|loading state
The current structure of fields is fetched.

While the data is fetched, a loading state is displayed.
If no fields, an empty state is displayed.
If an error is returned by the API, an error state is displayed.
If eveything's gonna be alright (don't you worry 🎶) then a basic list
of fields is displayed.

You can test it at: `/plugins/tracker/fields/<tracker-id>/`

Part of request #39759: Get Rid of Flaming Parrot in Trackers

Change-Id: Ie3782f0b1e933937ca2d24ef743fa20a6bc6b19e