chore: Remove indirect usages of argparse
We do not need it, we do not use the CLI of the dependencies
relying on it.

Overrided the deps tree with @types/argparse because our current
version of pnpm does not support the removal of deps in an
override.

No functional changes.

Part request #45595 Allowlist licenses acceptable for Tuleap dependencies

Change-Id: I50fb6dc948bac11c096418fec3011822255001fd