feat: Follow tracker structure
Previous iteration displayed all fields as a flat list, but we have
containers (fieldsets, columns).

Now we display fieldsets as tlp-pane, and columns as dashed div.

Having columns as real columns (vertically next to each other) will be
part of a dedicated contribution.

Part of request #39759: Get Rid of Flaming Parrot in Trackers

Change-Id: I22fae07112685b0d7dbce46fea1ed069c8c713c8