fix: Drop "tuleap-mercure" package
Mercure and Caddy are licensed under AGPL 3.0 making it
potentially complex to handle in corporate environments.

As a short term solution we just drop the package and reset
the feature flags. Longer term a minimal implementation of the
specification could be written.

Part request #45595 Allowlist licenses acceptable for Tuleap dependencies

Change-Id: I75c653067c770e19437aacf0d24ac1f55dbf5f70