feat: Display reviewers table
Part of request #39756 Get rid of Flaming Parrot in Document

Display all the reviewers of the approval table in a table. All the rows
are displayed in italic except for current user. For this one it's
review state is a link to the old review ui

If there is no reviewers an empty state is displayed

When fetching a single item, its approval table is retrieved without the
reviewers. So in CurrentApprovalTable we fetch the api to retrieve only
the approval table of the item with its reviewers and use it directly
instead of using the one in item

As the directory begin to contains a lot of files, I choose to move all
"display" component in a sub folder. This way later I'll be able to have
"creation", "admin" and "review" sub folders

*Testing:*

Note: You can force API return in DocmanItemsResource line 395

You should see the list of reviewers of the table

Change-Id: Ide96cc2c11d43b15bfc31d63bde441932bd654ca