feat: Shareable versions urls
part of story #44470 Pseudo-versions based on the artifact history

Whenever a user clicks on a version in the sidebar, a new state is
pushed into the browser history. This state contains the version id,
and allows us to:
1. Add/remove the version id from the query string of the current url
2. Be able to display the right version when users navigate
   backward/forward.

Note:

The fake versions are not shareable beacause we need to retrieve the
desired version from the backend beforehand using the version id found
in the query string.

How to test:
- Go to an artidoc containing artifact sections
- Press the [v] key
- Display the artifacts versions
--> `?version=<version_id>` is added to the current url when the
displayed version is not the latest

- Go backward/forward using the navigation arrows of your browser (or
  using alt+left / alt+right keys combinations)
--> The right versions is displayed

- Copy the url of a given version
- Paste it in another tab
--> The right version is displayed
--> The version tab in the sidebar is displayed

Change-Id: Ifdf368f68b480285729fde2870e4091eb7022ed8