fix: request #45640 Artifact link field not displayed after transition error
The div.tracker_hidden_edition_field was wrongly used for link field. I
don't know why it was used at first but the script on artifact view
remove fields with it.

Note: There is still one problem which need some work on the links
component -> if you have added a link from the artifact, in the error
view the link is not shown. This is because the component does not take
care of the initial value and simply fetch artifact current links.

*Testing:*

Update an artifact without respecting transition rules, the artifact
links field should be displayed

Change-Id: Ic87f0a6db69d336b0bea6f355f25785a3d1f4ef0