feat: Display previous tables in readonly
Part of request #39756 Get rid of Flaming Parrot in Document

First column of history table now have links to see previous table
state. When in this view the reviewer table is in readonly mode and you
cannot change your own review.

You can come back to current table by clicking on its row or on the tab

*Testing:*

When clicking on a previous table, all reviewers should be in italic and
your review column should not be clickable

Change-Id: If2638eed20ac39a2858cde08735fad60bb5a520b